+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,16 +1,20 @@
 🔀 What does this PR do? & Why are we doing this?
+
 <!-- 무슨 작업했는지 요약 -->
 
--[ ] icons 페이지에 크기 조절 size 추가 등
+- [ ] icons 페이지에 크기 조절 size 추가 등
 
 ---
 
 📸 Screenshots
+
 <!-- 참고 스크린샷 이미지 넣기 -->
 
 ---
 
 📌 Related Issue
+
 <!-- 관련된 이슈 번호가 있다면 써주기 -->
 <!-- Reference 버튼 누르면 번호 볼 수 있습니다~ -->
+
 Closes #번호
