+++ src/components/gauge/EarningsCardMonth.tsx
