+++ src/components/gauge/EarningsCardToday.tsx
