+++ src/components/gauge/GaugeComponent.tsx
