+++ src/components/gauge/GaugeComponentContainer.tsx
