+++ src/components/gauge/GaugeViewToggle.tsx
