+++ src/components/gauge/Header.tsx
