+++ src/components/gauge/WonSigns.tsx
