+++ src/components/input/WageInput.tsx
