+++ src/pages/Step1InputPage.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import InputPageLayout from '../components/InputPageLayout';
+import InputPageLayout from '../components/input/InputPageLayout';
+import WageInput from '../components/input/WageInput';
+import WageTypeToggle from '../components/input/WageTypeToggle';
 import InputPageModal from '../components/InputPageModal';
-import WageInput from '../components/WageInput';
-import WageTypeToggle from '../components/WageTypeToggle';
 import { useWage } from '../contexts/WageContext';
 
 function Step1InputPage() {
